+++ apps/dash-svm/assets/github/images/animated1.gif
