+++ apps/dash-svm/assets/github/images/screenshot.png
