+++ apps/dash-svm/assets/images/dash-logo-new.png
