+++ apps/dash-svm/requirements.txt
@@ -1,10 +1,5 @@
-# Core
-gunicorn>=19.8.1
-dash>=1.0.0
-
-# Additional
-colorlover>=0.2.1
-numpy>=1.16.2
-pandas>=0.24.2
-scikit-learn>=0.20.3
-scipy>=1.2.1
+dash==2.4.1
+pandas==1.4.2
+scikit-learn==1.1.1
+colorlover==0.3.0
+gunicorn==20.1.0
