+++ apps/dash-svm/utils/helper_functions.py
@@ -1,0 +1,32 @@
+from sklearn import datasets
+import numpy as np
+
+def generate_data(n_samples, dataset, noise):
+    if dataset == "moons":
+        return datasets.make_moons(n_samples=n_samples, noise=noise, random_state=0)
+
+    elif dataset == "circles":
+        return datasets.make_circles(
+            n_samples=n_samples, noise=noise, factor=0.5, random_state=1
+        )
+
+    elif dataset == "linear":
+        X, y = datasets.make_classification(
+            n_samples=n_samples,
+            n_features=2,
+            n_redundant=0,
+            n_informative=2,
+            random_state=2,
+            n_clusters_per_class=1,
+        )
+
+        rng = np.random.RandomState(2)
+        X += noise * rng.uniform(size=X.shape)
+        linearly_separable = (X, y)
+
+        return linearly_separable
+
+    else:
+        raise ValueError(
+            "Data type incorrectly specified. Please choose an existing dataset."
+        )
